requirements.txt
dnslib
publicsuffix2
cfpy
